requirements.txt
aioredis==2.0.1
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
certifi==2025.1.31
chardet==3.0.4
click==8.1.8
colorama==0.4.6
databases==0.9.0
exceptiongroup==1.2.2
fastapi==0.115.8
fastapi-cache2==0.2.2
googletrans
greenlet==3.1.1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httpx==0.27.2
hyperframe==5.2.0
idna==2.10
iniconfig==2.0.0
Jinja2==3.1.5
Mako==1.3.8
MarkupSafe==3.0.2
packaging==24.2
pendulum==3.0.0
pluggy==1.5.0
psycopg2==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
redis==5.2.1
rfc3986==1.5.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.45.3
time-machine==2.16.0
tomli==2.2.1
typing_extensions==4.12.2
tzdata==2025.1
uvicorn==0.34.0
